+++ src/js/videojs.record.js
@@ -50,7 +50,7 @@ class Record extends Plugin {
     /**
      * The constructor function for the class.
      *
-     * @param {(videojs.Player|Object)} player
+     * @param {(videojs.Player|Object)} player - video.js Player object.
      * @param {Object} options - Player options.
      */
     constructor(player, options) {
@@ -117,7 +117,7 @@ class Record extends Plugin {
     /**
      * Setup plugin options.
      *
-     * @param {Object} options - Optional new player options.
+     * @param {Object} newOptions - Optional new player options.
      */
     loadOptions(newOptions = {}) {
         let recordOptions = videojs.mergeOptions(pluginDefaultOptions,
@@ -130,6 +130,7 @@ class Record extends Plugin {
         this.recordAnimation = recordOptions.animation;
         this.recordScreen = recordOptions.screen;
         this.maxLength = recordOptions.maxLength;
+        this.maxFileSize = recordOptions.maxFileSize;
         this.msDisplayMax = parseFloat(recordOptions.msDisplayMax);
         this.debug = recordOptions.debug;
         this.recordTimeSlice = recordOptions.timeSlice;
@@ -155,6 +156,7 @@ class Record extends Plugin {
         this.audioBitRate = recordOptions.audioBitRate;
         this.audioChannels = recordOptions.audioChannels;
         this.audioMimeType = recordOptions.audioMimeType;
+        this.audioBufferUpdate = recordOptions.audioBufferUpdate;
 
         // animation settings
         this.animationFrameRate = recordOptions.animationFrameRate;
@@ -328,6 +330,21 @@ class Record extends Plugin {
                 this.surfer.liveMode = true;
                 this.surfer.surfer.microphone.paused = false;
 
+                // assign custom reloadBufferFunction for microphone plugin to
+                // obtain AudioBuffer chunks
+                if (this.audioBufferUpdate === true) {
+                    this.surfer.surfer.microphone.reloadBufferFunction = (event) => {
+                        if (!this.surfer.surfer.microphone.paused) {
+                            // redraw
+                            this.surfer.surfer.empty();
+                            this.surfer.surfer.loadDecodedBuffer(event.inputBuffer);
+
+                            // store data and notify others
+                            this.player.recordedData = event.inputBuffer;
+                            this.player.trigger('audioBufferUpdate');
+                        }
+                    };
+                }
                 // open browser device selection dialog
                 this.surfer.surfer.microphone.start();
                 break;
@@ -405,8 +422,9 @@ class Record extends Plugin {
 
     /**
      * Invoked when the device is ready.
+     *
      * @private
-     * @param stream: LocalMediaStream instance.
+     * @param {LocalMediaStream} stream - Local media stream from device.
      */
     onDeviceReady(stream) {
         this._deviceActive = true;
@@ -441,8 +459,8 @@ class Record extends Plugin {
                     ' is only supported in audio-only mode.');
             }
 
-            // determine recorder class
-            var EngineClass;
+            // get recorder class
+            let EngineClass;
             switch (this.audioEngine) {
                 case RECORDRTC:
                     // RecordRTC.js (default)
@@ -479,7 +497,6 @@ class Record extends Plugin {
                 // connect stream to recording engine
                 this.engine = new EngineClass(this.player, this.player.options_);
             } catch (err) {
-                console.error(err);
                 throw new Error('Could not load ' + this.audioEngine +
                     ' plugin');
             }
@@ -609,7 +626,9 @@ class Record extends Plugin {
 
     /**
      * Invoked when an device error occurred.
+     *
      * @private
+     * @param {(string|number)} code - Error code/description.
      */
     onDeviceError(code) {
         this._deviceActive = false;
@@ -941,9 +960,9 @@ class Record extends Plugin {
      */
     onCountDown() {
         if (!this.paused) {
-            var now = new Date().getTime();
-            var duration = this.maxLength;
-            var currentTime = (now - (this.startTime + this.pausedTime)) / 1000;
+            let now = new Date().getTime();
+            let duration = this.maxLength;
+            let currentTime = (now - (this.startTime + this.pausedTime)) / 1000;
 
             this.streamDuration = currentTime;
 
@@ -970,6 +989,8 @@ class Record extends Plugin {
      * Get the current time of the recorded stream during playback.
      *
      * Returns 0 if no recording is available (yet).
+     *
+     * @returns {float} Current time of the recorded stream.
      */
     getCurrentTime() {
         let currentTime = isNaN(this.streamCurrentTime) ? 0 : this.streamCurrentTime;
@@ -1016,6 +1037,8 @@ class Record extends Plugin {
      * Get the length of the recorded stream in seconds.
      *
      * Returns 0 if no recording is available (yet).
+     *
+     * @returns {float} Duration of the recorded stream.
      */
     getDuration() {
         let duration = isNaN(this.streamDuration) ? 0 : this.streamDuration;
@@ -1086,6 +1109,9 @@ class Record extends Plugin {
      *     blob(s) you want to save. File extensions are added automatically.
      *     For example: {'video': 'name-of-video-file'}. Supported keys are
      *     'audio', 'video' and 'gif'.
+     * @example
+     * // save video file as 'foo.webm'
+     * player.record().saveAs({'video': 'foo'});
      */
     saveAs(name) {
         if (this.engine && name !== undefined) {
@@ -1096,7 +1122,8 @@ class Record extends Plugin {
     /**
      * Destroy plugin only.
      *
-     * Use `destroy` to remove the plugin and the player.
+     * Use [destroy]{@link Record#destroy} to remove the plugin and the player
+     * as well.
      */
     dispose() {
         // disable common event listeners
@@ -1118,7 +1145,7 @@ class Record extends Plugin {
         this.player.clearInterval(this.countDown);
 
         // dispose wavesurfer.js
-        if (this.getRecordType() == AUDIO_ONLY) {
+        if (this.getRecordType() === AUDIO_ONLY) {
             if (this.surfer) {
                 // also disposes player
                 this.surfer.destroy();
@@ -1212,6 +1239,8 @@ class Record extends Plugin {
 
     /**
      * Mute LocalMediaStream audio and video tracks.
+     *
+     * @param {boolean} mute - Whether or not the mute the track(s).
      */
     muteTracks(mute) {
         if ((this.getRecordType() === AUDIO_ONLY ||
@@ -1228,6 +1257,10 @@ class Record extends Plugin {
 
     /**
      * Get recorder type.
+     *
+     * @returns {string} Recorder type constant.
+     * @example
+     * console.log(player.record().getRecordType()); // 'audio_video'
      */
     getRecordType() {
         return getRecorderMode(this.recordImage, this.recordAudio,
@@ -1271,10 +1304,11 @@ class Record extends Plugin {
     /**
      * Capture frame from camera and copy data to canvas.
      * @private
+     * @returns {void}
      */
     captureFrame() {
-        var detected = detectBrowser();
-        var recordCanvas = this.player.recordCanvas.el().firstChild;
+        let detected = detectBrowser();
+        let recordCanvas = this.player.recordCanvas.el().firstChild;
 
         // set the canvas size to the dimensions of the camera,
         // which also wipes the content of the canvas
@@ -1292,8 +1326,8 @@ class Record extends Plugin {
             if ((detected.browser === 'chrome' && detected.version >= 60) &&
                (typeof ImageCapture === typeof Function)) {
                 try {
-                    var track = this.stream.getVideoTracks()[0];
-                    var imageCapture = new ImageCapture(track);
+                    let track = this.stream.getVideoTracks()[0];
+                    let imageCapture = new ImageCapture(track);
                     // take picture
                     imageCapture.grabFrame().then((imageBitmap) => {
                         // get a frame and copy it onto the canvas
@@ -1319,6 +1353,8 @@ class Record extends Plugin {
     /**
      * Draw image frame on canvas element.
      * @private
+     * @param {HTMLCanvasElement} canvas - Canvas to draw on.
+     * @param {HTMLElement} element - Element to draw onto the canvas.
      */
     drawCanvas(canvas, element) {
         canvas.getContext('2d').drawImage(
@@ -1355,7 +1391,7 @@ class Record extends Plugin {
      * @private
      */
     showAnimation() {
-        var animationDisplay = this.player.animationDisplay.el().firstChild;
+        let animationDisplay = this.player.animationDisplay.el().firstChild;
 
         // set the image size to the dimensions of the recorded animation
         animationDisplay.width = this.player.width();
@@ -1393,13 +1429,15 @@ class Record extends Plugin {
     /**
      * Received new timestamp (when timeSlice option is enabled).
      * @private
+     * @param {float} current - Current timestamp.
+     * @param {array} all - List of timestamps so far.
      */
     onTimeStamp(current, all) {
         this.player.currentTimestamp = current;
         this.player.allTimestamps = all;
 
         // get blob (only for MediaStreamRecorder)
-        var internal;
+        let internal;
         switch (this.getRecordType()) {
             case AUDIO_ONLY:
                 internal = this.engine.engine.audioRecorder;
@@ -1412,24 +1450,40 @@ class Record extends Plugin {
             default:
                 internal = this.engine.engine.videoRecorder;
         }
-        internal = internal.getInternalRecorder();
+
+        let maxFileSizeReached = false;
+        if (internal) {
+            internal = internal.getInternalRecorder();
+        }
+
         if ((internal instanceof MediaStreamRecorder) === true) {
             this.player.recordedData = internal.getArrayOfBlobs();
 
             // inject file info for newest blob
             this.engine.addFileInfo(
                 this.player.recordedData[this.player.recordedData.length - 1]);
+
+            // check max file size
+            if (this.maxFileSize > 0) {
+                let currentSize = new Blob(this.player.recordedData).size;
+                if (currentSize >= this.maxFileSize) {
+                    maxFileSizeReached = true;
+                }
+            }
         }
 
         // notify others
         this.player.trigger('timestamp');
+
+        // automatically stop when max file size was reached
+        if (maxFileSizeReached) {
+            this.stop();
+        }
     }
 
     /**
      * Collects information about the media input and output devices
      * available on the system.
-     *
-     * Returns an array.
      */
     enumerateDevices() {
         if (!navigator.mediaDevices || !navigator.mediaDevices.enumerateDevices) {