+++ CHANGES.md
@@ -12,8 +12,14 @@ x.x.x - unreleased
 ------------------
 
 - Add support for screen capture (#289)
+- Add `maxFileSize` option to limit the maximum file size of a recorded
+  clip, and stop recording when that limit is reached (#234)
 - Add `msDisplayMax` option to control display of time format (#188)
 - Documentation: add Angular and Vue.js wiki pages (#274, #283)
+- Bump required videojs-wavesurfer version to 2.6.2 for microphone support
+  in the MS Edge and Safari browsers (#294)
+- Fix issue with `timeSlice` option when resetting the player (#300 by @GDIBass)
+- Examples: add Safari/Edge browser workarounds for audio-only recording (#295)
 
 **Backwards-incompatible changes** (when upgrading from a previous version):